requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pytest==8.4.1
httpx==0.28.1
google-generativeai==0.8.3
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
